requirements.txt
django>=1.10,<1.11
html2text
markdown
mysqlclient>=1.3.3
Pillow
pycrypto
pytz
requests
